+++ app/src/main/java/com/tworld/tviewing/MainActivity.kt
@@ -1,14 +1,14 @@
-package com.example.tviewing
+package com.tworld.tviewing
 
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import com.example.tviewing.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
     }
-
 
 
 }
