+++ app/src/main/res/layout/fragment_search.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#111111"
-    tools:context=".search.SearchFragment">
+    tools:context="com.tworld.tviewing.search.SearchFragment">
 
 
     <RelativeLayout
