+++ app/src/main/res/layout/fragment_video_detail.xml
@@ -4,10 +4,12 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".videoDetail.VideoDetailFragment">
+    tools:context="com.tworld.tviewing.videoDetail.VideoDetailFragment">
+
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent">
+
         <VideoView
             android:id="@+id/detail_videoView"
             android:layout_width="match_parent"
@@ -17,33 +19,36 @@
             app:layout_constraintTop_toTopOf="parent" />
 
         <LinearLayout
-            android:orientation="vertical"
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            android:padding="10dp"
             android:background="#656565"
+            android:orientation="vertical"
+            android:padding="10dp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/detail_videoView">
+
             <LinearLayout
-                android:orientation="horizontal"
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content">
+                android:layout_height="wrap_content"
+                android:orientation="horizontal">
+
                 <TextView
                     android:id="@+id/detail_title"
                     android:layout_width="0dp"
                     android:layout_height="wrap_content"
                     android:layout_weight="8"
-                    android:textSize="30sp"
                     android:text="Title"
-                    android:textColor="#DCDDDC"/>
+                    android:textColor="#DCDDDC"
+                    android:textSize="30sp" />
+
                 <TextView
                     android:id="@+id/detail_hits"
-                    android:text="조회수 : 12,345"
-                    android:textColor="#DCDDDC"
                     android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"/>
+                    android:layout_height="wrap_content"
+                    android:text="조회수 : 12,345"
+                    android:textColor="#DCDDDC" />
             </LinearLayout>
 
 
@@ -52,10 +57,12 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:orientation="horizontal">
+
                 <LinearLayout
-                    android:orientation="vertical"
                     android:layout_width="wrap_content"
-                    android:layout_height="wrap_content">
+                    android:layout_height="wrap_content"
+                    android:orientation="vertical">
+
                     <ImageButton
                         android:id="@+id/detail_isLike"
                         android:layout_width="30dp"
@@ -65,17 +72,20 @@
                         android:scaleType="fitCenter"
                         android:src="@drawable/ic_heart"
                         app:tint="#DCDDDC" />
+
                     <TextView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
                         android:layout_gravity="center"
                         android:text="찜"
-                        android:textColor="#DCDDDC"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"/>
+                        android:textColor="#DCDDDC" />
                 </LinearLayout>
+
                 <LinearLayout
-                    android:orientation="vertical"
                     android:layout_width="wrap_content"
-                    android:layout_height="wrap_content">
+                    android:layout_height="wrap_content"
+                    android:orientation="vertical">
+
                     <ImageButton
                         android:id="@+id/detail_notify"
                         android:layout_width="30dp"
@@ -85,17 +95,20 @@
                         android:scaleType="fitCenter"
                         android:src="@drawable/ic_bell"
                         app:tint="#DCDDDC" />
+
                     <TextView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
                         android:layout_gravity="center"
                         android:text="알림"
-                        android:textColor="#DCDDDC"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"/>
+                        android:textColor="#DCDDDC" />
                 </LinearLayout>
+
                 <LinearLayout
-                    android:orientation="vertical"
                     android:layout_width="wrap_content"
-                    android:layout_height="wrap_content">
+                    android:layout_height="wrap_content"
+                    android:orientation="vertical">
+
                     <ImageButton
                         android:id="@+id/detail_share"
                         android:layout_width="30dp"
@@ -105,12 +118,13 @@
                         android:scaleType="fitCenter"
                         android:src="@drawable/ic_share"
                         app:tint="#DCDDDC" />
+
                     <TextView
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
                         android:layout_gravity="center"
                         android:text="찜"
-                        android:textColor="#DCDDDC"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"/>
+                        android:textColor="#DCDDDC" />
                 </LinearLayout>
             </LinearLayout>
 
@@ -118,14 +132,14 @@
                 android:layout_width="match_parent"
                 android:layout_height="2dp"
                 android:layout_marginVertical="10dp"
-                android:background="#DCDDDC"/>
+                android:background="#DCDDDC" />
 
             <TextView
                 android:id="@+id/detail_content"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:text="content"
-                android:textColor="#DCDDDC"/>
+                android:textColor="#DCDDDC" />
         </LinearLayout>
     </androidx.constraintlayout.widget.ConstraintLayout>
 </FrameLayout>
