+++ app/src/test/java/com/tworld/tviewing/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
-package com.example.tviewing
+package com.tworld.tviewing
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
